data explorer: `polars` convert to code backend (#9077)
Added new base classes `FilterHandler` and `SortHandler` with subclasses
for polars and pandas.

Implemented and tested similarly to the `pandas` backend #8716

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #8753 Add **Convert to Code** support for `polars` data frames.

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer

I've added a handful of unit tests here to check that the dataframe
generated from the "convert to code" is equivalent to an expected
output. I've also updated the e2e test to check polars dataframes as
well.

You can do a quick overview by using the flights data in
qa-example-content. Unit tests cover all the operations, although not
for all types.

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>