Data Explorer: Request profiles one at a time for larger tables (#9096)
These are some changes to make profile rendering a little smoother.
Before we were batching 4 profiles at a time, but this changes to
1-at-a-time for over 1 million rows (below that they the in-window
profiles are requested in a single big batch):

I also noticed that the background task queue in Python was using 5 x #
CPUs by default and for compute heavy profile computation this can
result in inefficient context switching when computing many profiles.
Now the number of workers is equal to the number of logical CPU cores.