Reduce console error/warning noise (#9270)
This change removes a good chunk (but not all) of the noisy errors and
warnings in the console output that happen even in a perfectly normal
Positron startup.

Addresses #9182.


## API Proposals

The biggest offender; VS Code has strict handling for API proposals and
ensures that an extension explicitly declare exactly which proposals it
wants; this must match the set of proposals enabled for the extension on
the VS Code side. Because Positron doesn't enforce API proposals, we
have allowed the API proposal sets to get very stale, resulting in many
errors like these:

<img width="1025" height="505" alt="image"
src="https://github.com/user-attachments/assets/8e5264d3-162e-4bf8-ac54-1081c4873ad7"
/>

The fix is to turn off more parts of the API proposal validation system,
since we don't use it in Positron.

## Can't read properties of `experiment`

This was introduced with the upstream 103 merge. On every boot, this
shows up in the console:

<img width="1023" height="259" alt="image"
src="https://github.com/user-attachments/assets/7811127b-6db1-4e79-a38c-6f59db87c757"
/>

It appears that this is caused by experimental setting schema missing.
There's a recent 103 change to tolerate this, but it doesn't handle the
case where the schema is missing entirely.

## Can't parse Positron version

This has been in builds for some time:

<img width="798" height="144" alt="image"
src="https://github.com/user-attachments/assets/791903af-c42a-4c6b-bb1c-d96143066a2a"
/>

The problem is just that this code is sometimes called when we don't
have a version _at all_ (that is, version is just an empty string),
which is really fine when we're doing e.g. an update check and there is
no previous version.

The fix is to allow the codepaths that parse the version to return
`undefined` (i.e. "no version") when given an empty string, and to
handle that value safely elsewhere.