Refactor parent folder update logic in `FolderNameLocationStep` component (#9329)
Addresses https://github.com/posit-dev/positron/issues/8905

The "Empty Folder" template does not have a `nextStep`, so it was
skipping the part of the logic that updated where the new folder gets
put.

@:new-folder-flow

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fixed bug in parent folder handling ("Location") for _Workspaces: New
Folder from Template..._ when using the "Empty Folder" template.


### QA Notes

- Run _Workspaces: New Folder from Template..._ and choose "Empty
Folder"
- Choose a parent folder ("Location") that is _not_ the default
- Confirm that the new folder is created where you specified (not the
default)
- Run _Workspaces: New Folder from Template..._ again and choose a
different option
- Again, choose a parent folder ("Location") that is _not_ the default
- Confirm that this new folder also is created where you specified