Assistant: Improve git integration logging and model selection (#9289)
This PR addresses both:
 * https://github.com/posit-dev/positron/issues/9287
 * https://github.com/posit-dev/positron/issues/9288

Changes:
* Logging to the Assistant output pane has been added during the git
commit message generation process.
* The last used model ID is tracked in the `ParticipantService` and used
to select which model is used when generating git commit messages. Note
that we do not have our usual access to a the chat request object here,
which is where we normally get the model ID.
 * The previous model selection behaviour is retained as a fallback.
* If something goes wrong, it is emitted in the logs, and shown in the
UI.
 
## QA

As mentioned in https://github.com/posit-dev/positron/issues/9287,
getting into a situation where a provider is broken can be tricky.
Expired credentials on bedrock is the easiest way.

Nevertheless, with this PR we should now observe:

* Success and failure logging when generating a git commit in the
Assistant logs.
* UI errors when generation fails.

To test that the selected model is updating:

1) Start a new Assistant chat message with a given provider/model, say
hello to the model.
2) Add some dummy changes to a workspace with a git repo and generate a
git commit message.
3) Ensure that the model in the git commit message log matches the last
used LLM provider.
4) Repeat from 1) with a different provider/model.

---



https://github.com/user-attachments/assets/793eb3d0-019a-4313-ae28-f8342b2879ab