Update null filter labels to match summary panel (#9269)
Addresses #9083

The `is null` and `is not null` filter labels in the Data Explorer have
been updated to now say `is missing` and `is not missing`. These labels
match the label we use in the summary panel for calculating the
percentage of values that are null/missing.

The change in labels should hopefully resolve the awkwardness R users
experience since `na` is the value being represented by that label.

  


https://github.com/user-attachments/assets/3647d507-f591-41b5-9170-6ad318798b19



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Update null filter labels to match summary panel (#9083)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
@:data-explorer @:web @:win