ci: bump setup-node (#9377)
### Summary

Bumping setup-node to v5.

For the `e2e-ubuntu`, `integration` and `unit` workflows I opted to keep
our custom caching, but for `e2e-windows`, since we don't have a cache
strategy at present day, enabled the default cache that is now part of
v5. let's see if there is a noticeable difference.


### QA Notes

@:win @:web

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>