data explorer: add + guard keybinding for convert-to-code (#9325)
We have to add _a_ keybinding in order to have the when clause be
evaluated, so I made it `Cmd/Ctrl+K C`, but no strong feelings about this
if folks like something else!

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- https://github.com/posit-dev/positron/issues/9278 Fixes bug where
keybindings can forcibly open Convert to Code modal


### QA Notes

- try `Cmd/Ctrl+K C` in a file, no Convert to Code modal
- try `Cmd/Ctrl+K C` in a data explorer with no filters or sorts, no
Convert to Code modal
- try `Cmd/Ctrl+K C` in a data explorer with filters or sorts, see Convert
to Code modal