Restore Copilot language server binaries (#9396)
The recent update to the Copilot LSP had the side effect of _removing_
the binary entirely due to some furniture rearrangement on Microsoft's
part -- the native binaries are no longer all installed by default;
instead each one is a separate npm package.

The annoying thing about this is that only the correct package for your
OS gets installed, but when doing cross-compilation we need to get
packages for other OS's. So now we need to force install a couple of
packages on macOS to pick up both architectures.

Addresses https://github.com/posit-dev/positron/issues/9384.