e2e: add support to run playwright tests against external server (#9400)
### Summary

Adds the ability to run tests against external server.
1. Start server: `npm run e2e-start-server`
2. Select project: `e2e-browser-external`
3. Run tests the way you normally do...

Note: there seems to be some bugs related to sessions (sessions
persisting and not deleting), I added some logic to handle those a bit
better, but it's still a little buggy. I'll be opening ticket for that
shortly.

### QA Notes

Not enabling these in CI or anything yet, just getting things ready.