Positron Assistant LLM testing initial. (#9418)
This implements the first real LLM/inspect-ai testing for Positron
assistant. It's mainly the framework with a couple basic tests.

For now, it's skipped until I get the details for running in a CI
workflow worked out (the placeholder for that is part of this PR).

- `README.md` covers the basics for the testing. This is really a draft
and will be updated in subsequent PRs
- There are some wait's and console.log's sprinkled about for various
debugging, etc. These are temporary and will be removed when it's
running in some CI.
- The playwright test is currently skipped to ensure it doesn't run in
other CI right now. Once a separate workflow is running, this will be
cleaned up as well.

### QA Notes
If you want to try it out, follow the instructions in the `README.md`.
_note:_ you'll need to unskip the playwright file.
<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->