Fix obscured total token usage (#9417)
This moves the total token usage to above the files changed part. The
files changed was designed to be adjacent to the rest of the input part
so it makes sense to move the token usage up. Added another fix to
adjust the spacing below the input part. It seems that there's negative
margins on the files changed part that affects the spacing below the
chat input area that needs to account for when the token usage is
visible or not.

<img width="396" height="219" alt="image"
src="https://github.com/user-attachments/assets/b2336dc0-6f94-42ad-a35b-f0909e6b7ded"
/>


### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix obscured total token usage in Assistant


### QA Notes