Add support for pinning in summary panel (#9220)
Addresses #9196, #8570

When a column is pinned in the data grid, the corresponding row for that
column also gets pinned in the summary panel. The order of the pinned
columns will match the order in the data grid. For example, if columns
1, 9, 6 are pinned (in that order), then the pinned rows in the summary
panel should also be 1, 9, 6. Pinned rows will not show up in the
un-pinned rows section. When a column is un-pinned, the summary panel
will immediately unpin the corresponding row as well.

When searching/sorting in the summary panel, the pinned rows should
always be present at the top, with the search results listed below. If a
pinned row would be in the search results, it should instead not show up
twice in the dataset but only show up once as a pinned row.

This PR also adds some missing keyboard support for the summary panel.
You can now use `tab` and `shift+tab` to navigate through the rows.
Pressing enter when focused on a row should allow you to expand/collapse
the details for that row.


https://github.com/user-attachments/assets/2b0ab111-cee0-4b15-ac68-326ddc371f4b


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:data-explorer @:web @:win

---------

Co-authored-by: Brian Lambert <brianlambert@gmail.com>