Positron Notebooks: Fix kernel selection (#9461)
This fixes kernel selection when:

1. Saving an untitled notebook
(https://github.com/posit-dev/positron/issues/9021)
2. Opening a new notebook - we now autoselect the kernel matching the
current globally selected runtime using the same logic as done in
https://github.com/posit-dev/positron/pull/2498.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

Save an untitled notebook and ensure that the kernel is not lost.

Open a new notebook while Python is the globally selected interpreter
and check that a Python kernel starts.

@:notebooks