Add inspect-ai test workflow (#9495)
Add a workflow for the inspect-ai LLM testing.

Test is currently skipped as we are going to move things around a bit, but wanted to get this into `main` before making that change.

Adds some new things in teh tests to allow for overriding input and outputs. And fixed a utf-8 processing error that could crop up depending on the model's response.


### QA Notes

<!--
  Positron team members: please add relevant e2e test tags, so the tests can be
  run when you open this pull request.

  - Instructions: https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
  - Available tags: https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->